requirement-Explore Advanced Retrievers in LlamaIndex.txt
llama-index==0.12.49
llama-index-embeddings-huggingface==0.5.5
llama-index-llms-ibm==0.4.0
llama-index-retrievers-bm25==0.5.2
sentence-transformers==5.0.0
rank-bm25==0.2.2
PyStemmer==2.2.0.3
langchain_openai